various changes

+++ MassDelete.py
@@ -1,22 +1,12 @@
 import glob
 import os
+import regex
 
-files = glob.glob("/Users/michaelingram/PycharmProjects/TPC_1099/*public.xml")
+keyword =''
+files = glob.glob("/Volumes/SSD/production/*/*.csv")
 for x in files:
-    #print(x)
-    os.remove(x)
-
-
-
-import os
-import sys
-
-os.chdir('/Users/michaelingram/.Trash')
-if len(sys.argv) >= 2:
-    if sys.argv[1] == '-t' or sys.argv[1] == '-T':
-        os.system("tree ./")
-    elif sys.argv[1] == '-l' or sys.argv[1] == '-L':
-        os.system("ls -al")
-else:
-    print("Nothing in the bin to delete")
-os.system("rm -rf *")
+    if os.path.exists(x):
+        os.remove(x)
+    else:
+        #print('No file found')
+        pass
